Fix false positives, use global_prios.csv and wrap tweeting in function
