+++ global_prios/global_prios.csv
@@ -4,4 +4,13 @@ Cure Cancer,1e8,(cancer OR tumor OR tumour) AND cure ,https://80000hours.org/art
 US land use,1e6,US AND "land use",https://80000hours.org/articles/problem-framework/
 US teacher red tape,1e5,(US AND bureaucracy AND teachers),https://80000hours.org/articles/problem-framework/
 Mental health, 1.7e8, (schizophrenia OR depression OR anxiety), global burden of disease study
-Smoking, 1e8, (smoking OR cigarette) AND (health OR death OR cancer OR stroke OR "heart attack"), https://www.thelancet.com/journals/lancet/article/PIIS0140-6736(17)30819-X/fulltext 
+Smoking, 1e8, (smoking OR cigarette) AND (health OR death OR cancer OR stroke OR "heart attack"), https://www.thelancet.com/journals/lancet/article/PIIS0140-6736(17)30819-X/fulltext
+Climate Change, 5.5e6, "climate change", https://www.who.int/globalchange/environment/en/chapter7.pdf
+Air pollution, 7e6, "air pollution", https://www.who.int/air-pollution/news-and-events/how-air-pollution-is-destroying-our-health
+heart disease,1.8e8, "heart disease", PMID_30415748
+TB,160000000.0,tuberculosis,PMID_30415748
+pneumonia,110000000,pneumonia,PMID_30415748
+diarrhoea,81000000,diarrhoea,PMID_30415748
+neonatal_preterm,70000000,neonatal AND death,PMID_30415748
+HIV,54000000,HIV,PMID_30415748
+asthma,23000000,asthma,PMID_30415748
